added proxy option
This tries to download the required files from the server but if you are behind a corporate proxy server it will fail. By using the environment variables to check for a proxy setting this could be avoided.
Also this will fix all `imagemin` modules like `imagemin-optipng` and `imagemin-pngquant` that deep down are trying to download the compiled files for windows.