Add the ability to accelerate the download by dividing the file into segments and run concurrent connections.
1. minSegmentSize: number (default 265 KB)
2. maxConnectionsCount: number (default 4, number of concurrent connection)
3. fileExistAction: rename | overwrite | skip (default overwrite)
4. accelerate: true | false (default true, download by concurrent connections)

fix problem when downloading large files, since they get downloaded to the memory. The files new are downloaded to the disk directly. When no destination is provided, the default OS temp folder is used.

Since the file is downloaded by segments, this will lay out the bases to provide resumable file download. The feature is not implemented yet.

The files list now contains, objects that have url and location (file path in the disk)

Known issues:
1. Since the file is downloaded in segments, this confuses download status because it will show progress bar for each segment